moved over flag processing functions
