+++ src/Helpers/EmpireHelper.ts
@@ -16,196 +16,6 @@ import RoomVisualHelper from "Managers/RoomVisuals/RoomVisualHelper";
 export default class EmpireHelper {
 
     /**
-     * commit a remote flag to memory
-     * @param flag the flag we want to commit
-     */
-    public static processNewRemoteFlag(flag: Flag): void {
-
-        // Get the host room and set the flags memory
-        const dependentRoom: Room = Game.rooms[this.findDependentRoom(flag.pos.roomName)];
-        const flagTypeConst: FlagTypeConstant | undefined = this.getFlagType(flag);
-        const roomName: string = flag.pos.roomName;
-        Memory.flags[flag.name].complete = false;
-        Memory.flags[flag.name].processed = true;
-        Memory.flags[flag.name].timePlaced = Game.time;
-        Memory.flags[flag.name].flagType = flagTypeConst;
-        Memory.flags[flag.name].flagName = flag.name;
-        Memory.flags[flag.name].spawnProcessed = false;
-
-        // Create the RemoteFlagMemory object for this flag
-        const remoteFlagMemory: RemoteFlagMemory = {
-            flagName: flag.name,
-            flagType: flagTypeConst
-        };
-
-
-        // If the dependent room already has this room covered, set the flag to be deleted and throw a warning
-        const existingDepedentRemoteRoomMem: RemoteRoomMemory | undefined = _.find(MemoryApi.getRemoteRooms(dependentRoom),
-            (rr: RemoteRoomMemory) => {
-                if (rr) {
-                    return rr.roomName === roomName;
-                }
-                return false;
-            });
-
-        if (existingDepedentRemoteRoomMem) {
-            Memory.flags[flag.name].complete = true;
-            throw new UserException(
-                "Already working this dependent room!",
-                "The room you placed the remote flag in is already being worked by " + existingDepedentRemoteRoomMem.roomName,
-                ERROR_WARN);
-        }
-
-        // Otherwise, add a brand new memory structure onto it
-        const remoteRoomMemory: RemoteRoomMemory = {
-            sources: { cache: Game.time, data: 1 },
-            hostiles: { cache: Game.time, data: null },
-            structures: { cache: Game.time, data: null },
-            roomName: flag.pos.roomName,
-            flags: [remoteFlagMemory],
-            reserveTTL: 0,
-        };
-
-        MemoryApi.createEmpireAlertNode("Remote Flag [" + flag.name + "] processed. Host Room: [" + dependentRoom.name + "]", 10);
-        dependentRoom.memory.remoteRooms!.push(remoteRoomMemory);
-    }
-
-    /**
-     * commit a attack flag to memory
-     * @param flag the flag we want to commit
-     */
-    public static processNewAttackFlag(flag: Flag): void {
-
-        // Get the host room and set the flags memory
-        const dependentRoom: Room = Game.rooms[this.findDependentRoom(flag.pos.roomName)];
-        const flagTypeConst: FlagTypeConstant | undefined = this.getFlagType(flag);
-        const roomName: string = flag.pos.roomName;
-        Memory.flags[flag.name].complete = false;
-        Memory.flags[flag.name].processed = true;
-        Memory.flags[flag.name].timePlaced = Game.time;
-        Memory.flags[flag.name].flagType = flagTypeConst;
-        Memory.flags[flag.name].flagName = flag.name;
-        Memory.flags[flag.name].spawnProcessed = false;
-
-        // Create the RemoteFlagMemory object for this flag
-        const attackFlagMemory: AttackFlagMemory = this.generateAttackFlagOptions(flag, flagTypeConst, dependentRoom.name);
-
-
-        // If the dependent room already has this room covered, just push this flag onto the existing structure
-        const existingDepedentAttackRoomMem: AttackRoomMemory | undefined = _.find(MemoryApi.getAttackRooms(dependentRoom),
-            (rr: AttackRoomMemory) => {
-                if (rr) {
-                    return rr.roomName === roomName;
-                }
-                return false;
-            });
-
-        if (existingDepedentAttackRoomMem) {
-            MemoryApi.createEmpireAlertNode("Attack Flag [" + flag.name + "] processed. Added to existing Host Room: [" + existingDepedentAttackRoomMem.roomName + "]", 10);
-            existingDepedentAttackRoomMem.flags.push(attackFlagMemory);
-            return;
-        }
-
-        // Otherwise, add a brand new memory structure onto it
-        const attackRoomMemory: AttackRoomMemory = {
-            hostiles: { cache: Game.time, data: null },
-            structures: { cache: Game.time, data: null },
-            roomName: flag.pos.roomName,
-            flags: [attackFlagMemory],
-        };
-
-        MemoryApi.createEmpireAlertNode("Attack Flag [" + flag.name + "] processed. Host Room: [" + dependentRoom.name + "]", 10);
-        dependentRoom.memory.attackRooms!.push(attackRoomMemory);
-    }
-
-    /**
-     * commit a claim flag to memory
-     * @param flag the flag we want to commit
-     */
-    public static processNewClaimFlag(flag: Flag): void {
-
-        // Get the host room and set the flags memory
-        const dependentRoom: Room = Game.rooms[this.findDependentRoom(flag.pos.roomName)];
-        const flagTypeConst: FlagTypeConstant | undefined = this.getFlagType(flag);
-        const roomName: string = flag.pos.roomName;
-        Memory.flags[flag.name].complete = false;
-        Memory.flags[flag.name].processed = true;
-        Memory.flags[flag.name].timePlaced = Game.time;
-        Memory.flags[flag.name].flagType = flagTypeConst;
-        Memory.flags[flag.name].flagName = flag.name;
-        Memory.flags[flag.name].spawnProcessed = false;
-
-        // Create the ClaimFlagMemory object for this flag
-        const claimFlagMemory: ClaimFlagMemory = {
-            flagName: flag.name,
-            flagType: flagTypeConst
-        };
-
-
-        // If the dependent room already has this room covered, set the flag to be deleted and throw a warning
-        const existingDepedentClaimRoomMem: ClaimRoomMemory | undefined = _.find(MemoryApi.getClaimRooms(dependentRoom),
-            (rr: ClaimRoomMemory) => {
-                if (rr) {
-                    return rr.roomName === roomName;
-                }
-                return false;
-            });
-
-        if (existingDepedentClaimRoomMem) {
-            Memory.flags[flag.name].complete = true;
-            throw new UserException(
-                "Already working this dependent room!",
-                "The room you placed the claim flag in is already being worked by " + existingDepedentClaimRoomMem.roomName,
-                ERROR_WARN);
-        }
-
-        // Otherwise, add a brand new memory structure onto it
-        const claimRoomMemory: ClaimRoomMemory = {
-            roomName: flag.pos.roomName,
-            flags: [claimFlagMemory],
-        };
-
-        MemoryApi.createEmpireAlertNode("Claim Flag [" + flag.name + "] processed. Host Room: [" + dependentRoom.name + "]", 10);
-        dependentRoom.memory.claimRooms!.push(claimRoomMemory);
-    }
-
-    /**
-     * commit a depedent room over-ride flag to memory
-     * @param flag the flag we are commiting to memory
-     */
-    public static processNewDependentRoomOverrideFlag(flag: Flag): void {
-
-        // Set all the memory values for the flag
-        const flagTypeConst: FlagTypeConstant | undefined = this.getFlagType(flag);
-        Memory.flags[flag.name].complete = false;
-        Memory.flags[flag.name].processed = true;
-        Memory.flags[flag.name].timePlaced = Game.time;
-        Memory.flags[flag.name].flagType = flagTypeConst;
-        Memory.flags[flag.name].flagName = flag.name;
-        Memory.flags[flag.name].spawnProcessed = false;
-
-        MemoryApi.createEmpireAlertNode("Option Flag [" + flag.name + "] processed. Flag Type: [" + RoomVisualHelper.convertFlagTypeToString(flagTypeConst) + "]", 10);
-    }
-
-    /**
-     * commit a stimulate flag to an owned room
-     * @param flag the flag we are commiting to memory
-     */
-    public static processNewStimulateFlag(flag: Flag): void {
-
-        // Set all the memory values for the flag
-        const flagTypeConst: FlagTypeConstant | undefined = this.getFlagType(flag);
-        Memory.flags[flag.name].complete = false;
-        Memory.flags[flag.name].processed = true;
-        Memory.flags[flag.name].timePlaced = Game.time;
-        Memory.flags[flag.name].flagType = flagTypeConst;
-        Memory.flags[flag.name].flagName = flag.name;
-        Memory.flags[flag.name].spawnProcessed = false;
-
-        MemoryApi.createEmpireAlertNode("Option Flag [" + flag.name + "] processed. Flag Type: [" + RoomVisualHelper.convertFlagTypeToString(flagTypeConst) + "]", 10);
-    }
-
-    /**
      * finds the closest colonized room to support a
      * Remote/Attack/Claim room
      * Calls helper functions to decide auto or over-ride
