+++ src/Helpers/ProcessFlagHelpers/ProcessStimulateFlag.ts
@@ -1,3 +1,7 @@
+import EmpireHelper from "Helpers/EmpireHelper";
+import RoomVisualHelper from "Managers/RoomVisuals/RoomVisualHelper";
+import MemoryApi from "Api/Memory.Api";
+
 export class ProcessDefaultStimulateFlag implements IFlagProcesser {
 
     public primaryColor: ColorConstant = COLOR_GREEN;
@@ -9,10 +13,18 @@ export class ProcessDefaultStimulateFlag implements IFlagProcesser {
     }
 
     /**
-     * Process the default remote room flag
+     * Process the default stimulate flag
      * @param flag
      */
     public processFlag(flag: Flag): void {
+        const flagTypeConst: FlagTypeConstant | undefined = EmpireHelper.getFlagType(flag);
+        Memory.flags[flag.name].complete = false;
+        Memory.flags[flag.name].processed = true;
+        Memory.flags[flag.name].timePlaced = Game.time;
+        Memory.flags[flag.name].flagType = flagTypeConst;
+        Memory.flags[flag.name].flagName = flag.name;
+        Memory.flags[flag.name].spawnProcessed = false;
 
+        MemoryApi.createEmpireAlertNode("Option Flag [" + flag.name + "] processed. Flag Type: [" + RoomVisualHelper.convertFlagTypeToString(flagTypeConst) + "]", 10);
     }
 }
